[CryptAuth v2] Pass ClientAppMetadata directly instead of a provider
In the previous CL [2/4], we fetched ClientAppMetadata as part of the
DeviceSync service initialization. In this CL, we pass the
ClientAppMetadata directly to the classes that need it. This eliminates
the technical complexity caused by the ClientAppMetadata provider's
asynchronous API.

---

This is ClientAppMetadata change-detection CL [3/4]:

This chain of CLs relates to the end goal of re-enrolling if
ClientAppMetadata has changed since the last successful enrollment. For
example, when Phone Hub become supported, we will want to re-enroll.

We were previously only fetching ClientAppMetadata on demand, for
instance, right before a periodic enrollment. Now, we need to fetch the
data every time the service starts up to see if it changed. We continue
to make the assumption that the data will not change during a user
session.

Bug: 1105703
Change-Id: I41920152a299a7137d48acd7755718f023b1957b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2319329
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Commit-Queue: Josh Nohle <nohle@chromium.org>
Cr-Commit-Position: refs/heads/master@{#793699}