Simplified domain: set elided parts of the URL to transparent
When the simplified domain field trials are enabled, ElideURL() is
sometimes used to hide most of the URL except for (part of) the
domain. This is implemented by using GetSubstringBounds() to find the
portion of the string that should not be elided, and setting the
display rect and offset to show only that portion. However,
GetSubstringBounds() can return bounds that just slightly larger (+/-
1 pixel) than the actual string, due to internal rounding. This can
result in small artifacts around the simplified domain, as shown in
the screenshots in the bug.

To hide these artifacts, this CL sets the hidden portions of the URL
to transparent in ElideURL(). The hidden portions are already set to
transparent when we use ElideAnimation to animate from full URL to
simplified domain, because fading is part of the animation. So in this
CL we only need to add transparency in ElideURL() which is called when
an animation isn't necessary.

Bug: 1108045
Change-Id: Ie146b3585a908067a722dfcec6b6e0c9d03071d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2323721
Commit-Queue: Emily Stark <estark@chromium.org>
Reviewed-by: Justin Donnelly <jdonnelly@chromium.org>
Cr-Commit-Position: refs/heads/master@{#793708}