[CryptAuth v2] Re-enroll if ClientAppMetadata changes
We store a hash of the last successfully enrolled ClientAppMetadata. On
startup, if we detect that the metadata has changed since the last
successful enrollment, we force a re-enrollment.

Note: We do *not* force an enrollment if the persisted hash has not
been set. So, a user will need to enroll before this check takes
effect.

---

This is ClientAppMetadata change-detection CL [4/4]:

This chain of CLs relates to the end goal of re-enrolling if
ClientAppMetadata has changed since the last successful enrollment. For
example, when Phone Hub become supported, we will want to re-enroll.

We were previously only fetching ClientAppMetadata on demand, for
instance, right before a periodic enrollment. Now, we need to fetch the
data every time the service starts up to see if it changed. We continue
to make the assumption that the data will not change during a user
session.

Bug: 1105703
Change-Id: I6bb868e325498c91f90a3965fb31e3237eb0df65
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2300879
Commit-Queue: Josh Nohle <nohle@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#793726}