Have the main histogram owner be a person w/an @google/@chromium email.
For histograms without an appropriate primary owner, add either the
placeholder text or add an appropriate individual.

Note that this is enforced for only non-obsolete histograms.

While this should encourage people adding new histograms to have an
appropriate primary owner, it is still technically possible for them
to list the placeholder text as the owner to get around this. I think
this is okay because reviewers would presumably not approve this.

Bug: b/147376752, b/162326579, b/160411669
Change-Id: I6a3df3d8326528572a4606a1baa0c2ddc8ea7f34
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2329737
Commit-Queue: Caitlin Fischer <caitlinfischer@google.com>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Cr-Commit-Position: refs/heads/master@{#793732}