v8 code cache: web_test with HTTP cache revalidation
The test loads a resource that goes through revalidation each time and
returns with status 304 (Not Modified). The test checks that
revalidation causes flushing of hte code cache on disk.

Whether this is intentional behavior or not, I am not sure. Just wanted
to document the current behavior and compare it to cache discarding
later.

Bug: 1045052
Change-Id: I629ae3789811cfce53558a13795060b6c6895a76
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2327492
Commit-Queue: Egor Pasko <pasko@chromium.org>
Reviewed-by: Mythri Alle <mythria@chromium.org>
Cr-Commit-Position: refs/heads/master@{#793742}