Sync dependency visualization URL with page state
The page's URL now automatically updates itself to match the page state
whenever the page state is changed. Also, more of the page's state is
stored in the URL - previously it was only the node filter, now the
following graph settings are stored as well:

* Inbound/outbound depth
* Display settings (curved edges, coloring on hover, color schemes)
* Convex hull display (for class graph only)

As a result of the change, the URL generator component is now redundant.
It has been removed.

Bug: 1106944
Change-Id: I617640d7648b8083dbdff41522f1c1f914496fc4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2324491
Commit-Queue: James Long <yjlong@google.com>
Reviewed-by: Mohamed Heikal <mheikal@chromium.org>
Reviewed-by: Samuel Huang <huangs@chromium.org>
Cr-Commit-Position: refs/heads/master@{#793752}