Return OAuth token scopes in the GetAuthToken API
To add unbundled consent for GetAuthToken in the extensions API, the
token scopes returned from the server needs to be accessible for the
extension to know which scopes were approved or rejected. This CL adds
the functionality and API changes to allow extensions to obtain the
list of accepted scopes. Until unbundled consent is implemented, these
returned scopes do not add any additional useful information to the
GetAuthToken function.

Bug: 1100535
Change-Id: Iba3233ae8a33c49118e7093fd2e9020b2951575f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2303596
Commit-Queue: William Lin <williamlin@google.com>
Reviewed-by: David Roger <droger@chromium.org>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#793777}