Handle errors in GetCodedSizeAndVisibleRect
In case of ImageReader the UpdateTexImage can fail even if
RenderToTextureOwnerFrontBuffer succeeded. This can lead to situation
when we don't have current frame to get size from. This CL handles
this situation to fallback to visible size until we will be able to
get right size.

Bug: 1109016
Change-Id: I2fc6381cccc4560e97355ac465f62a72734facef
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2328013
Reviewed-by: Frank Liberato <liberato@chromium.org>
Reviewed-by: Peng Huang <penghuang@chromium.org>
Commit-Queue: Vasiliy Telezhnikov <vasilyt@chromium.org>
Cr-Commit-Position: refs/heads/master@{#793789}