WebLayer: attempt to de-flake FullscreenSizeTest
Dismiss the OSK before entering fullscreen. Otherwise, the tap that
enters fullscreen also dismisses the OSK, and there could be a race
between how those two actions affect the page height.

I was unable to reproduce the flake so this is a speculative fix.

Bug: 1107651
Change-Id: I92e3bb502305fa9468e5639d6411f6a6ecb0ba90
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2333058
Reviewed-by: Haiyang Pan <hypan@google.com>
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Evan Stade <estade@chromium.org>
Cr-Commit-Position: refs/heads/master@{#793814}