Add Observing windows when modify AppService instance for M85.
Sometimes, during the system init phases,the windows are not added to
the observe list in AppServiceAppWindowLauncherController, so when the
window is destroyed, AppServiceAppWindowLauncherController is not
notified, and the window is not removed from AppService instance, which
might cause crash when operating the destroyed window saved in instance.

This CL modified the code to check and add the window to
AppServiceAppWindowLauncherController's |observed_windows_|, when the
window is added/modified in AppService instance, so that the window
destroy callback in AppServiceAppWindowLauncherController can catch the
window destroy information to remove the window recorder.

BUG=1011235

TBR=xiyuan@chromium.org

(cherry picked from commit d8ee8b15baf5470c5f190d2648459f2d4e1bcd92)

Change-Id: I48ee1a7fc55adc9097a53f18149f44672ac10a92
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2307669
Commit-Queue: Nancy Wang <nancylingwang@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#790712}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2332070
Reviewed-by: Nancy Wang <nancylingwang@chromium.org>
Cr-Commit-Position: refs/branch-heads/4183@{#1130}
Cr-Branched-From: 740e9e8a40505392ba5c8e022a8024b3d018ca65-refs/heads/master@{#782793}