Android: Set the incognito profile destruction right
It is possible on Android platform that more than one ProfileDestroyer
instance exists, targeted to a single Profile. UAF crash happens when
the second one comes along to destroy the profile which was already
destroyed by the first one. This CL ensures that the remaining
Destroyers get their targeted profile reset to null as the first one
destroys it to avoid such incidence.

The issue was already handled by https://crrev.com/c/1971031 but got
reverted due to a side effect. This is another attempt but in much
more simpler way.

Bug: 1029677
Change-Id: Icff3d08c8fa94050f2ed6bee740157542474735e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2320393
Reviewed-by: Mihai Sardarescu <msarda@chromium.org>
Reviewed-by: Ramin Halavati <rhalavati@chromium.org>
Commit-Queue: Jinsuk Kim <jinsukkim@chromium.org>
Cr-Commit-Position: refs/heads/master@{#793954}