[WebSocket] Remove DOMArrayBuffer from the send queue
This commit removes DOMArrayBuffer from the WebSocket send queue, and
replaces it with unique_ptr<char*>. The original implementation creates
a new DOMArrayBuffer to queue the message. This requires GC to run
redundant times.

To prevent this, change DOMArrayBuffer to use dynamically allocated
char*, and use unique_ptr deleter to remove when the queue is processed.

Bug: 987883
Change-Id: I6bd20dcc002a4e04f75f3b38ca2a8d86835a7832
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2280809
Commit-Queue: Yutaka Hirano <yhirano@chromium.org>
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Reviewed-by: Adam Rice <ricea@chromium.org>
Cr-Commit-Position: refs/heads/master@{#794052}