[Chrome Cleaner] Replace base::string16 with std::wstring 3/4
This change is part of a series of changes that replacs usages of
base::string16 with std::wstring in chrome/chrome_cleaner and fixes
appropriate includes. This CL fixes the following subdirectories:

* chrome/chrome_cleaner/os
* chrome/chrome_cleaner/parsers

This change is a simple search and replace and does not effect any
functionality.

Bug: 911896
Change-Id: Ic9b497f6179621988996b9d59a68f6ebf489020e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2333826
Reviewed-by: proberge <proberge@chromium.org>
Commit-Queue: Jan Wilken Dörrie <jdoerrie@chromium.org>
Cr-Commit-Position: refs/heads/master@{#794111}