[Signin][Android] Collapse to the added account for web sign-in flow
This CL adjusts the add-account flow in web sign-in bottom sheet with
multiple accounts on device. Prior to this CL, we returned to the
expanded account list with the added account once it is added.
With this CL, we return to the collapsed account list with the
added account once the account in question is added to the device.

Bug: 1097031
Change-Id: I5842d59b365d9901798ad5177d80296db42135fd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2332601
Commit-Queue: Alice Wang <aliceywang@chromium.org>
Reviewed-by: Tanmoy Mollik <triploblastic@chromium.org>
Cr-Commit-Position: refs/heads/master@{#794119}