DLP: Add RestrictionSet instead of boolean confidentiality flag.
As there might be multiple different restrictions applied to the content,
having a single boolean "confidentiality" flag is not enough and the state
should be presented by a set of current restrictions.
This CL introduced DlpContentRestrictionSet class that uses bitmask to
store the restriction set (single restriction for now, but will be expanded)
and migrates other classes to use it instead of bool flag.

Bug: 1104884
Test: Unit tests updated.
Change-Id: Id157192b1f2b58b7eff71d6fb6c4cfc915cee8d0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2318145
Commit-Queue: Sergey Poromov <poromov@chromium.org>
Reviewed-by: Aya Elsayed <ayaelattar@chromium.org>
Reviewed-by: Nikita Podguzov <nikitapodguzov@chromium.org>
Cr-Commit-Position: refs/heads/master@{#794120}