[identity] Do not propagate token changes to getAuthToken on Desktop
Updating cookies from
GaiaRemoteConsentFlow::OnEndBatchOfRefreshTokenStateChanges() is
needed on ChromeOS to push new accounts added from the account manager.

On Desktop, sign-in happens on the Web and a new account is directly
added to getAuthToken's partition cookie jar. An extra update triggered
from OnEndBatchOfRefreshTokenStateChanges might change cookies order in
the middle of the consent flow and lead to a bug like
https://crbug.com/1112343.

This CL makes the cookie update in OnEndBatchOfRefreshTokenStateChanges
executed only on ChromeOS.

Bug: 1112343
Change-Id: Ica8c888a3572f406b8c282c17049ba9306bf7a4f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2332192
Reviewed-by: David Roger <droger@chromium.org>
Commit-Queue: Alex Ilin <alexilin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#794149}