VideoToolbox/SharedImage: Add signal to ignore IOSurfaceIsInUse
Note that IOSurfaceIsInUse is accessed via the function
GLImage::IsInUseByWindowServer. This patch description uses the
IOSurfaceIsInUse nomenclature.

Prior to this patch, a GLImageIOSurface would ignore IOSurfaceIsInUse
if it was initialized by specifying a CVPixelBuffer. This is because
the IOSurfaces that come from CVPixelBuffer that come from VideoToolbox
always return true for IOSurfaceIsInUse.

In subsequent patches, we will no longer be directly creating a
GLImageIOSurface from a CVPixelBuffer, but rather will be going through
the SharedImage interface. Add a mechanism though which these paths can
indicate that IOSurfaceIsInUse cannot be trusted, by adding a flag
SHARED_IMAGE_USAGE_MACOS_VIDEO_TOOLBOX, and a helper function
GLImage::DisableInUseByWindowServer.

Bug: 1108909
Change-Id: I06cd6d70d2700080e426844dec30bb8cff13b77d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2333053
Reviewed-by: Geoff Lang <geofflang@chromium.org>
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Commit-Queue: ccameron <ccameron@chromium.org>
Cr-Commit-Position: refs/heads/master@{#794186}