[Offline indicator v2] Fix toolbar swipe switch to native pages
BrowserControlsManager#setTab() restores controls positions for the tab
the user switches to. This caused a bug with the status indicator where
the native page would have its old controls offsets restored even after
the status indicator is shown.

This CL prevents the situation above by showing the controls when the
controls are restored for a native page.

Bug: 1111202
Change-Id: I75bb1f9d9851800f123bc6bab87307a191696202
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2332567
Commit-Queue: Sinan Sahin <sinansahin@google.com>
Reviewed-by: Theresa  <twellington@chromium.org>
Reviewed-by: Jinsuk Kim <jinsukkim@chromium.org>
Cr-Commit-Position: refs/heads/master@{#794220}