Disable Autofill when lastTappedWindow != keyWindow
Due to a bug in iOS 13 multiwindow system, touch events randomly do not
set the keyWindow. The result is that you can interact with a page and
continue to tap in another page.
This is specially bad for autofill, where it becomes possible to fill
private information in the wrong page.
This CL keep track of the last window that has been tapped and disables
autofill if this is not the keywindow.

Bug: 1109124
Change-Id: I6ef0797ca74f425999fe30e607541bc6757ffcb3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2317345
Commit-Queue: Olivier Robin <olivierrobin@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Reviewed-by: Javier Ernesto Flores Robles <javierrobles@chromium.org>
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#794227}