[iOS CBCM] Add support for CloudPolicyOverridesPlatformPolicy
Adds support for the CloudPolicyOverridesPlatformPolicy policy, which
controls whether policies fetched from CBCM should take precedence over
policies found on the device.

This policy is a requirement for launching CBCM. For now, it is
declared as a Future policy for iOS until iOS CBCM is launched.

Bug: 1066495
Change-Id: I5b2ce20e2f2b435992f84747a47a3c2ea0f14273
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2332322
Reviewed-by: Owen Min <zmin@chromium.org>
Commit-Queue: Guillaume Jenkins <gujen@google.com>
Cr-Commit-Position: refs/heads/master@{#794233}