Add gradient mask to omnibox elide animation
Adds a rectangle that gradients from omnibox background color to
transparent on the left side of the omnibox, and an inverted one to
the right.
The rectangles are used as gradient masks when an elide animation is
running, to smooth out the color transition at the end.
The mask is fixed size, except if they would obscure the not-elided part
of the URL, in which case they shrink.

Screencast of change (public link): https://docs.google.com/presentation/d/1AGXJNkt8_oQyPLSDOfAGduscL0LRCCUcZq5XtJo2eC0/edit?usp=sharing

Bug: 1106467
Change-Id: Ic5029bbc850285ebe4e5e76a5a3d810c965f7e44
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2321883
Commit-Queue: Carlos IL <carlosil@chromium.org>
Reviewed-by: Tommy Li <tommycli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#794235}