Set CFBundleName for SxS Chromes
In the past, this wasn't done for Canary because the idea was for
it to be as identical to the normal Chrome Stable as possible. Now
that we're making Chromes to run side-by-side to Chrome Stable,
they need their own identity lest visual issues like the one in
the linked bug appear.

This relands commits 21e6c4e122543f5ab2b1996d4423c05568b150be and
34b007d8a255723c93fba527071be732fee2805e. They were reverted
because they broke crash uploading; the change in this CL of
crashpad_mac.mm addresses that.

Bug: 1109411
Change-Id: Ib85d773a8ade79422826fe99d88254cd40a35b0e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2333074
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Reviewed-by: Mark Mentovai <mark@chromium.org>
Commit-Queue: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#794267}