service worker: Return live version reference from GetRegistrationsForOrigin()
The method is used by ServiceWorkerRegistry and it creates a
ServiceWorkerVersion in DidGetRegistrationsForOrigins(). The method
should pass a ServiceWorkerLiveVersionRef so that a ServiceWorkerVersion
can keep it during the lifetime of the ServiceWorkerVersion.

This CL:
* Merges SerializedServiceWorkerRegistration into
  ServiceWorkerFindRegistrationResult as these are almost the same.
* Moves |status| out of ServiceWorkerFindRegistrationResult because
  it doesn't much make sense for GetRegistrationsForOrigin()
* Tweaks ServiceWorkerStorageControlImpl to create
  ServiceWorkerLiveVersionRef for GetRegistrationsForOrigin()

This CL doesn't pass a live version reference to ServiceWorkerVersion
yet. A follow-up CL will pass the reference to it.

Bug: 1055677
Change-Id: I408f741e7f6a199fdc162f4e68c98233b2b6a50a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2331961
Commit-Queue: Kenichi Ishibashi <bashi@chromium.org>
Reviewed-by: Makoto Shimazu <shimazu@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#794335}