VideoToolbox/SharedImage: CreateImageForGpuMemoryBuffer from Mach port
GpuMemoryBufferFactoryIOSurface::CreateImageForGpuMemoryBuffer currently
ignores the Mach port specified in the GpuMemoryBufferHandle, and rather
relies on the GpuMemoryBufferId and client_id.

Allow using the Mach port if the GpuMemoryBufferId is unspecified. Add
the restriction that this only work if the client_id indicates that this
is a reserved in-process client (like the video decoder), because (at
present) only in-process clients will need this path. This restriction
can be lifted in the future if there is need.

Bug: 1108909
Change-Id: I261c80494f66666c776d9e0d4dd96af524598392
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2332849
Commit-Queue: ccameron <ccameron@chromium.org>
Reviewed-by: Daniele Castagna <dcastagna@chromium.org>
Cr-Commit-Position: refs/heads/master@{#794379}