appcache: fix data race in unit tests
Service worker reads the scoped feature list via leveldb's ChromiumEnv.
Many AppCacheUpdateJobTests write a scoped feature list too, and this is
a data race.  Fix this race by moving the setting of the scoped feature
list to the constructor of each test class and moving the instantiation
of the BrowserContext to the SetUp function.

Bug: 1111119
Change-Id: If5b4adbf495a841e22bbee5fe17f4870ee42e866
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2335820
Reviewed-by: Marijn Kruisselbrink <mek@chromium.org>
Commit-Queue: enne <enne@chromium.org>
Cr-Commit-Position: refs/heads/master@{#794398}