CrOS: Fix icon standardizer to detect smaller circle icons.
Without this CL, the IsIconCircleShaped() helper function would only
detect a circle shape if the icon took up the maximum space of its
bitmap.

By temporarily scaling the visible icon to the bounds of its bitmap,
this CL allows IsIconCircleShaped() to detect when the icon is circle
shaped. This means a circular icon will be detected even when the
visible icon does not take up the width and height of the entire bitmap.

Also in this change, the icon_standardizer helper functions loop through
all of the ImageSkiaReps instead of looking at the ImageSkiaRep which
has the max supported scale.

Bug: 1110496
Change-Id: Ia59906dd3b1ec1be41d1a7d165c6f032813d7209
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2333336
Reviewed-by: Nancy Wang <nancylingwang@chromium.org>
Commit-Queue: Matthew Mourgos <mmourgos@chromium.org>
Cr-Commit-Position: refs/heads/master@{#794426}