PDF Viewer: Fix documentNeedsScrollbars logic after recent Viewport changes.
As of r791686, a |document.documentElement| instance (which is the
<html> DOM node) is passed to the Viewport constructor, instead of
|window|. Moreover |window.innerWidth| was replaced with
|document.documentElement.offsetLeft|. The latter does not include the
scrollbar width, unlike the former, which in turn caused the previous
logic in documentNeedsScrollbars() to not be 100% correct.

Scrollbar width no longer needs to be manually accounted for, and the
logic is fixed while simplified at the same time.

Bug: 1111525
Change-Id: I1a83d73e88a5c8f095b500c75dfec12ec1475d62
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2332049
Reviewed-by: Rebekah Potter <rbpotter@chromium.org>
Commit-Queue: dpapad <dpapad@chromium.org>
Cr-Commit-Position: refs/heads/master@{#794450}