[PwdCheckAndroid] Move the C++ bridge in chrome/browser
All the C++ files for this feature will need to be able to depend
on chrome/browser, because they require a Profile and a suite of
factories that reside in chrome/browser.

Due to the C++ part being instantiated from java, the necessary objects
cannot be injected.

Bug: 1102025, 1092444
Change-Id: Iebd6c283c0ba3305225ddefb1b0a12530cadd9a2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2332658
Reviewed-by: Friedrich [CET] <fhorschig@chromium.org>
Commit-Queue: Ioana Pandele <ioanap@chromium.org>
Cr-Commit-Position: refs/heads/master@{#794454}