[Autofill][Slimshady] Merging of structured components.
The merge logic supports combining observations in the scenario that
both root nodes have token equivalent values.

There are some other edge cases that we could support for v1 without
the need for additional UI, but this covers the scope we discussed in
the feature sync.

Change-Id: I6c773eeda6be328e50f1c7818b46585281f72d79
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2316200
Commit-Queue: Matthias Körber <koerber@google.com>
Reviewed-by: Dominic Battré <battre@chromium.org>
Cr-Commit-Position: refs/heads/master@{#794499}