Use early return instead of DCHECK when restoring cookies with CreateSantizedCookie.
For now CreateSanitizedCookie has a check where it returns a null
pointer when the cookie is marked secure but the URL has an http://
scheme. This check is in the process of being removed, but in the short
term we replace the DCHECK with an early return to prevent segfaults.

Bug: 1110788
Change-Id: Iff5da4de934fe762d2b3dd9ffc0a7d20c028f3a7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2335554
Commit-Queue: Dylan Cutler <dylancutler@google.com>
Reviewed-by: Lily Chen <chlily@chromium.org>
Reviewed-by: Tommy Nyquist <nyquist@chromium.org>
Cr-Commit-Position: refs/heads/master@{#794509}