Weaken time comparison in net_backoff_entry_serializer_fuzzer.
The fuzzer found a test case that causes the BackoffEntry deserializer
to discard the absolute timeout and fall back on the relative timeout.

BackoffEntry::GetReleaseTime() is actually not guaranteed to equal the
original release time (before serialization). However, it is guaranteed
to be less than or equal to the original release time.

This CL weakens the fuzzer's property check accordingly.

Bug: 1085995
Change-Id: I1dafb08575e73b0548f973565145c8ba7890a33e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2324111
Reviewed-by: Maksim Orlovich <morlovich@chromium.org>
Commit-Queue: Dan McArdle <dmcardle@chromium.org>
Cr-Commit-Position: refs/heads/master@{#794549}