media: Support MacOS ARM64 Widevine CDM as a component
This CL adds support for MacOS ARM64 Widevine CDM so that it will
be registered as a component.

Currently the CDM is not available on the server for component
update, not will it be bundled to Chrome. These will be resolved
later, as tracked by crbug/1111381.

Bug: 1110364,1111381
Change-Id: I4fa28d524f21bebdc13fbfb47ecdf8ece7b811ba
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2333997
Reviewed-by: John Rummell <jrummell@chromium.org>
Commit-Queue: Xiaohan Wang <xhwang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#794564}