Stops computing an artificial accessible name from inner text on Mac
When an object on Mac is focusable but has no value or name, we compute its accessible description
from its inner text.
Unfortunately, we were adding a space between the concatenation of each piece of text coming from each child object
which creates a mismatch between the text offsets returned by AXPosition and text markers vs. the actual text contents.
If extra spaces are needed, they should be added by Blink, not by the browser.
This patch removes the code that computes the inner text completely since
no other screen reader announces focusable items unless they have an accessible name,
and it is not proper to add an accessible name if there is none.

AX-Relnotes: On Mac and same as on all other platforms, navigating by tab now does not announce
the text contained in a large container such as a table row if that container is focusable.

R=dmazzoni@chromium.org

Change-Id: Iac5f321113595a2372df03bb119ab1c7a6f1bbf4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2308649
Commit-Queue: Nektarios Paisios <nektar@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#794574}