[PwdCheckAndroid] Save referrer in instance state
Although every password settings fragment should be launched using
PasswordManagerHelper#showPasswordSettings() which requires a referrer,
some clients crashed because the bundle seems to be missing regardless.

A possible reason is that the fragment wasn't started but just
recreated without a bundle. To cover these cases, store the referrer
in the savedInstanceState and reuse it on creation.

Bug: 1112628
Change-Id: I93805c2d93bae35c5dd0b6647626db87cc7db275
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2335444
Reviewed-by: Ioana Pandele <ioanap@chromium.org>
Commit-Queue: Friedrich [CET] <fhorschig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#794599}