Use only SafeSearch filter for EduCoexistence webview.
This cl fixes the bug where the child user is stuck
in EDU account addition flow when the user's parental
control setting is 'Only allow certain sites'.

Inorder to prevent this error from happening we will do
only safe search filter for the EduCoexistence webview
and not enforce the 'Only allow certain cites' settings
potentially set by the child's parents.

"https://accounts.google.com" is manually whitelisted. This
is because there is a race condition where we don't know
whether the WebContent is an inner WebContent belonging
to EduCoexistence flow or not by the time throttling happens.
This happens because navigation to accounts.google.com is
started before the <webview> WebContents is attached to
the outer WebContents hosted by the EduCoexistence flow WebUI.

We are making this exception only to the webview that is present in
the EDUCoexistence flow. The exception is made because the parent is
present with the child during the flow. Extending this exception
to every inner webcontents will may allow the child to break parental
settings.

Bug: 1079510
Change-Id: I4f3e65e58537a845613b0c49b03f801f0a3caa08
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2316443
Reviewed-by: Marc Treib <treib@chromium.org>
Reviewed-by: Aga Wronska <agawronska@chromium.org>
Commit-Queue: Yilkal Abe <yilkal@chromium.org>
Cr-Commit-Position: refs/heads/master@{#794661}