[Nearby] Add Nearby Share allowed policy
Adds the policy for allowing/disallowing Nearby Share via policy. A new
policy handler type is introduced to map a policy value of false to pref
value of false, but when the policy is true, the pref is not changed.
This allows the policy to force off a feature when disallowed, but the
user must still manually enable the feature if the policy allows it.
This avoids having to introduce another "allowed" pref.

For now, Nearby Share will only be supported on CrOS. However,
full desktop browser support is planned and this policy will be reused.
The policy is currently set to launch in M87 for CrOS so it is left with
no supported_on set yet.

Change-Id: Ibfcc4ececf49de646ef7c72872db9c6f14c68e97
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2316887
Commit-Queue: James Vecore <vecore@google.com>
Reviewed-by: Pavol Marko <pmarko@chromium.org>
Reviewed-by: Ryan Hansberry <hansberry@chromium.org>
Cr-Commit-Position: refs/heads/master@{#794682}