[omnibox] Basic implementation of on-clobber ZeroSuggest for Desktop
This CL is the most basic implementation possible of requesting
zero-prefix suggestions (ZeroSuggestProvider) when the user deletes the
permanent text in the omnibox.

In other words, when the flag is on, and the user enters
user_input_in_progress mode by deleting everything, this will request
ZeroSuggestions.

This doesn't implement the oft=2 mode yet, so there's no distinction
between on-clobber and on-focus requests.

Bug: 1106096
Change-Id: I2f0f3353605f82b378cc2d73499b8f03ceb83f32
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2335509
Reviewed-by: Justin Donnelly <jdonnelly@chromium.org>
Commit-Queue: Tommy Li <tommycli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#794691}