Clean up PlatformFontMacTest.DerivedFineGrainedFonts
Do comparisons on int-casted enums; it's easier to work out
what's going wrong if you don't have to convert from hex.

Use explicit capturing in the lambda; it's clearer than
the implicit reference capture.

This does not change the functionality of the test. It still
fails on Big Sur but it's more clear how it fails.

Bug: 1101426
Change-Id: I3a636129b62589b5cbccb5779c1e23920fc9da1b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2337106
Commit-Queue: Avi Drissman <avi@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Cr-Commit-Position: refs/heads/master@{#794710}