Allow alternative audio device authorization for GMC in RenderHostImpl.
The Global Media Controls (GMC) UI will allow the user to change the
audio output device used for a webpage. This change adds a method to
the RenderHostImpl allowing normal output device authorization to be
bypassed for a single device selected by the user.

The bypass works by storing a single device id hashed for a particular
security origin in the AudioOutputAuthorizationHandler. The result of
access permission checks will then be overwritten if the requested
device has the same id as what was stored.

Bug: 1105132
Change-Id: I9d22e617c502ed8d69fdddc06571017c7ba1f33e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2295824
Commit-Queue: Noah Rose Ledesma <noahrose@google.com>
Reviewed-by: Aaron Colwell <acolwell@chromium.org>
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Cr-Commit-Position: refs/heads/master@{#794722}