Report CRX signature verification time using UMA statistics
We want to improve the installation time for force installed
extensions. For this, we need to first gather the statistics regarding
which installation stages are time consuming. We collect the UMA
statistics for the time taken to verify the signature of CRX.

Following histogram is added:
Extensions.ForceInstalledTime.CRXDownloadCompleteTo.
VerificationComplete

Bug: 1102369
Change-Id: Ic1b38abcff8557fba1cd1c6b59de716d489355d6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2336080
Reviewed-by: Steven Holte <holte@chromium.org>
Reviewed-by: Oleg Davydov <burunduk@chromium.org>
Commit-Queue: Swapnil Gupta <swapnilgupta@google.com>
Cr-Commit-Position: refs/heads/master@{#794729}