[Shelf] Show navigation widget before setting the focus
The native view of the navigation widget is not activatable when its
target visibility is false. As a result, when the navigation widget
is hidden, it fails to get the focus through the FOCUS_SHELF
accelerator.

This CL fixes the issue by showing the navigation widget before setting
the focus.

Bug: 1111426
Change-Id: I9a1303ff6170eb98b255e6ce5d3814efce4d08c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2337097
Commit-Queue: Andrew Xu <andrewxu@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#794746}