Add 'max' attribute to month-picker-appearance-step.html to prevent 'This month' button from toggling
The month picker's 'This month' button is grayed out if the current
month (the actual month it is right now, not the 'value' attribute
of the control) is a valid choice as determined by the value, min,
max, and step attributes. August 2020 is the first month since this
test was introduced that is valid per the combination of value and
step of the control in this test, so it started failing at the
beginning of the month as the button was no longer grayed out.

This change re-enables the test, adding a 'max' attribute with a
value in the past, so that the button will always be grayed out
going forward.

The baselines still need to be regenerated because it turns out
that adding a max changes the width of the control's text field.
See blink::DateTimeNumericFieldElement::MaximumWidth and the
reference to it in
blink::DateTimeEditElement::CustomStyleForLayoutObject.

Bug: 1112111
Change-Id: I70cce1b98114c888a701e067d0543b7d12e2f457
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2337099
Reviewed-by: Mason Freed <masonfreed@chromium.org>
Commit-Queue: Dan Clark <daniec@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#794768}