GPU: Add SyncToken::ToDebugString()
This is intended as a convenience for debugging synchronization issues
related to sync tokens.

The output is in the format "a:b:c:d", i.e. "0:1:4:16", where a is
the namespace_id, b/c are the upper and lower half of the 64bit command
buffer ID (corresponding to channel and route ID for gpu/ipc/ usage),
and d is the release count.

Change-Id: I2cbc1ce9cff56d8b83bfa582ce4702f7459005c3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2324869
Commit-Queue: Kenneth Russell <kbr@chromium.org>
Auto-Submit: Klaus Weidner <klausw@chromium.org>
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Reviewed-by: Sunny Sachanandani <sunnyps@chromium.org>
Reviewed-by: Piotr Bialecki <bialpio@chromium.org>
Cr-Commit-Position: refs/heads/master@{#794776}