Introduce letter-spacing-override descriptor for @font-face (2/2)
Design doc: https://bit.ly/39qATQ4

This patch applies the letter-spacing-override value in @font-face,
and uses it to change the spacing between characters.

Note that the spacing is considered part of the font face, and hence,
it's applied at the character level instead of element. See the new
test case letter-spacing-override.html for an example.

Though we don't have a stable spec at the moment, we land
it behind a flag to make testing and discussion easier.

Bug: 1098355
Change-Id: I182d89aa9226ac92a0af50af06c0c56b021fb394
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2318510
Commit-Queue: Xiaocheng Hu <xiaochengh@chromium.org>
Reviewed-by: Dominik Röttsches <drott@chromium.org>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/master@{#794777}