Added save as IPC.
Due to some ux decisions the media app now needs to support a "save as"
operation where the current file is saved in a new location and the app
is navigated to the new file. This means that "saveCopy" is no longer an
accurate name so we are renaming it to "saveAs". Additionally this CL
adds some additional logic to slip the newly saved file into the
navigable files list so users can navigate back to the original file or
any other file in the directory.

To prevent TOT being broken this CL leaves saveCopy as is for older
version of the media app client code. Once the client is updated
saveCopy will be deleted in a separate CL.

Google3 side of this change: cl/323288209.

Bug: b/161193454
Change-Id: Ic66dcfdbcefbee985a9764701cb0bcee229020aa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2316189
Commit-Queue: Zain Afzal <zafzal@google.com>
Reviewed-by: Trent Apted <tapted@chromium.org>
Cr-Commit-Position: refs/heads/master@{#794829}