[directorytree] Turn selectByEntry() into an async function
selectByEntry() is not guaranteed to finish synchronously, however, the
callsite in onDirectoryChanged_() expects it to. This function became
async for two reasons:

1. The directory tree always used to have one more level preloaded, just
not visible. This meant that folder navigation on RHS was sync as it
could only traverse one level. This had changed in crrev.com/c/2059729,
and folder navigation is now async.
2. Navigations to deeply nested folders became possible using the "Go to
file location" feature.

Turn selectByEntry() into an async function to allow callers to await on
it. This fixes the active marker in the directory tree not being
updated. Also add a test to ensure the active marker updates correctly.

Bug: 1086370
Change-Id: I1423710bfd989bcb512f19ebca40e8e9bf405ad1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2331770
Commit-Queue: Austin Tankiang <austinct@chromium.org>
Reviewed-by: Noel Gordon <noel@chromium.org>
Reviewed-by: Luciano Pacheco <lucmult@chromium.org>
Cr-Commit-Position: refs/heads/master@{#794843}