Delete all entry points to the Genius app.
This removes the feature that it was conditioned on and adds logic in
extensions to remove the component and not add it again. We should be
able to remove it completely in another milestone after clients have had
a chance to remove it properly.

Bug: b/159863346
Change-Id: Id6369e670f749b6cd532a25eac0d35c718a4f5dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2318889
Commit-Queue: Rachel Carpenter <carpenterr@chromium.org>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: calamity <calamity@chromium.org>
Reviewed-by: Trent Apted <tapted@chromium.org>
Cr-Commit-Position: refs/heads/master@{#794857}