X11 and Ozone: Disable some parts of webinput unittests
These parts are not used in Ozone/Linux for the use_x11 && !use_ozone
config. However, we are going towards have a single Linux
config where OzonePlatform is enabled by
--enable-features=UseOzonePlatform. Thus, the parts that
are guarded by \#ifdef(USE_X11) start to be executed. This is not
desired for Ozone/Linux as it can run Wayland instead.

Thus, temporarily add check for the OzonePlatform feature flag and
a TODO to properly fix that.

-----

PS: Please note that this is a temp solution that will help to choose
between ozone and non-ozone X11 build. The switch that will be used
to choose the path is --enable-features=UseOzonePlatform. Once
non-Ozone X11 path is removed (hopefully by Q1 2021 depending on how
the finch trial goes), the wrapper will be removed.

Please also note that it's impossible to build use_x11 && use_ozone
without some hacks in PlatformCursor code. The changes to that are
on their way to upstream.

----

Bug: 1085700
Change-Id: I45cb0c9d36a7e45e72b8ff52cbfbdfcb37bc4ee5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2333199
Commit-Queue: Maksim Sisov (GMT+3) <msisov@igalia.com>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#794874}