X11 and Ozone: disable X11 specific tests in views_unittests
Enabling use_x11 && use_ozone results in running X11-only tests
that cannot succeed for Wayland config. Thus, disable them
(as it was before we started to work on use_x11 && use_ozone
config) via checking for the feature flag instead of relaying
on GN. I acknowledge that this is wrong to do, but I'm going to
fix that as soon as use_x11 && use_ozone config is upstreamed.

-----

PS: Please note that this is a temp solution that will help to choose
between ozone and non-ozone X11 build. The switch that will be used
to choose the path is --enable-features=UseOzonePlatform. Once
non-Ozone X11 path is removed (hopefully by Q1 2021 depending on how
the finch trial goes), the wrapper will be removed.

Please also note that it's impossible to build use_x11 && use_ozone
without some hacks in PlatformCursor code. The changes to that are
on their way to upstream.

----

Bug: 1085700
Change-Id: I77b943fe8e54982b97c40a4f34bbbca66ab3f726
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2336597
Reviewed-by: Thomas Anderson <thomasanderson@chromium.org>
Commit-Queue: Maksim Sisov (GMT+3) <msisov@igalia.com>
Cr-Commit-Position: refs/heads/master@{#794876}