Formalise that the AppServiceProxy doesn't exist in non-guest incognito profiles.
This CL removes a long-standing TODO, and adds DCHECKs in
AppServiceProxyFactory::GetForProfile that mirror the existing logic
used to determine whether to return a nullptr or a real AppServiceProxy.

Also fix a context menu test that was fetching an AppServiceProxy with
an incognito profile, and remove now-redundant DCHECKs/nullptr checks
that are placed immediately after a call to
AppServiceProxyFactory::GetForProfile that will now DCHECK if a nullptr
would be returned.

BUG=1112197

Change-Id: I9a7de8849f819cceb061677cef9e37c967d4651a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2334249
Commit-Queue: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Luciano Pacheco <lucmult@chromium.org>
Reviewed-by: Tony Yeoman <tby@chromium.org>
Reviewed-by: Aga Wronska <agawronska@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Reviewed-by: Trent Apted <tapted@chromium.org>
Reviewed-by: Nancy Wang <nancylingwang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#794897}