[iOS] Use the correct UserAgent during preload
When a page is loaded with the preloader, the correct user agent should
be used. When using AUTOMATIC user agent, the User Agent string is
determined by the trait collection of the WebView.
The code is using the container view of the WebView as a proxy for the
web view as the web view isn't always created when the user agent is
determined.

Currently as the preloader doesn't return a container view, the
default UIWindow screen is used.
It means that when the preloader is using multitasking, the view used
to determined the user agent is the whole screen of the device isn't of
being the window of the app, leading to a wrong user agent (as the
screen has a different trait collection than the window).

(cherry picked from commit 99f40819b9efa53b48670efebd06789d78ea8b64)

Fixed: 1100852
Change-Id: I874424e752e2f96e179f177a24b777112a4b5755
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2332665
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#794084}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2336725
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/branch-heads/4183@{#1225}
Cr-Branched-From: 740e9e8a40505392ba5c8e022a8024b3d018ca65-refs/heads/master@{#782793}