Add UMA to Chrome side for link to SW Internals page in DevTools
In order to avoid retrieving all service worker registrations in
DevTools, we removed the "other origins" UI[1] and added a link to
chrome://serviceworker-internals[2] if developers still want to see
registrations from other origins.

This UMA is added to check how many people are actually using that
link in the Application Panel in DevTools to view other origin
registrations. This is done by checking the query string of the url.

[1] https://crrev.com/c/2308615
[2] https://crrev.com/c/2324268

Bug: 807440
Change-Id: If6204792478352a65a0c30fa689d69ada7a00507
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2325537
Commit-Queue: Nidhi Jaju <nidhijaju@google.com>
Reviewed-by: Matt Falkenhagen <falken@chromium.org>
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Reviewed-by: Kenichi Ishibashi <bashi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#794931}